jsdb::database::table

Struct Table

pub struct Table<Value> {
    pub rows: HashMap<u32, Row<Value>>,
}

A table that organizes rows in a JSDB database.

The Table struct stores a collection of rows in a HashMap, where each row is identified by a unique u32 primary key. It provides methods to create a table, add rows, and remove rows, facilitating data organization within a database. The struct is generic, allowing flexible storage of various data types in its rows.

Type Parameters

Value: The type of data stored in the columns of the rows within this table.

Examples

use jsdb::JSDB;
use jsdb::database::row::Row;
use jsdb::database::table::Table;

// Create a JSDB environment and database
let mut project: JSDB<String> = JSDB::new();
let db = jsdb::database::database::Database::create_database();
project.add_database("my_db", db);

// Add a table and a row
if let Some(database) = project.databases.get_mut("my_db") {
    let table = Table::create_table();
    database.add_table("my_table", table);

    if let Some(table) = database.tables.get_mut("my_table") {
        let row = Row::create_row();
        table.add_row(1, row);
    }
}

Notes

Tables are stored in-memory and do not persist data to disk.

Row primary keys must be unique within a table to avoid overwriting.

Fields
rows: HashMap<u32, Row<Value>>

A collection of rows, mapped by their u32 primary keys.

Implementations

impl<Value> Table<Value>

pub fn create_table() -> Self

Creates a new, empty table.

Initializes a Table instance with an empty HashMap to store rows.

Returns

A new Table instance.

Examples

use jsdb::database::table::Table;
let table: Table<String> = Table::create_table();
assert!(table.rows.is_empty());

pub fn add_row(&mut self, primary_key: u32, row: Row<Value>)

Adds a row to the table.

Inserts a Row into the table’s HashMap with the specified u32 primary key. If a row with the same primary key already exists, it is replaced.

Arguments

primary_key: The unique u32 identifier for the row.

row: The Row instance to add.

Examples

use jsdb::database::table::Table;
use jsdb::database::row::Row;
let mut table: Table<String> = Table::create_table();
let row = Row::create_row();
table.add_row(1, row);
assert!(table.rows.contains_key(&1));

pub fn delete_row(&mut self, primary_key: u32)

Removes a row from the table.

Deletes the row with the specified u32 primary key. If no row exists with that primary key, the operation has no effect.

Arguments

primary_key: The u32 identifier of the row to remove.

Examples

use jsdb::database::table::Table;
use jsdb::database::row::Row;
let mut table: Table<String> = Table::create_table();
let row = Row::create_row();
table.add_row(1, row);
table.delete_row(1);
assert!(table.rows.is_empty());